changes to make type integer compatible with f2c

+++ ext/f2c_libs/f2c.h
@@ -23,8 +23,10 @@ typedef __int64 ulongint;  /* HACK ALERT */
 #pragma warning(disable:4101)
 #endif
 
-typedef long int integer;
-typedef unsigned long int uinteger;
+//typedef long int integer;
+typedef int integer;
+//typedef unsigned long int uinteger;
+typedef unsigned int uinteger;
 typedef char *address;
 typedef short int shortint;
 typedef float real;
@@ -58,9 +60,12 @@ typedef short flag;
 typedef short ftnlen;
 typedef short ftnint;
 #else
-typedef long int flag;
-typedef long int ftnlen;
-typedef long int ftnint;
+//typedef long int flag;
+//typedef long int ftnlen;
+//typedef long int ftnint;
+typedef int flag;
+typedef int ftnlen;
+typedef int ftnint;
 #endif
 
 /*external read, write*/
